refactor: extract worktree view/path leaf helpers
Move the stateless worktree-view mapping, lookup, branch-cleanup, and path
helpers out of the 1640-line service.go into worktree_helpers.go, leaving
service.go focused on the stateful service methods.
